- [ZBX-536] fixed problem with "Unable to add performance counter" (Sasha)  - [ZBX-535] added support of monitoring files larger than 2GB in windows (Sasha)
git-svn-id: svn://svn.zabbix.com/trunk@6122 97f52cf1-0a1b-0410-bd0e-c28be96e8082

+++ src/libs/zbxcommon/str.c
@@ -290,14 +290,14 @@ void __zbx_zbx_snprintf_alloc(char **str, int *alloc_len, int *offset, int max_l
  ******************************************************************************/
 void	zbx_strcpy_alloc(char **str, int *alloc_len, int *offset, const char *src)
 {
-	size_t	sz;
+	int	sz;
 
 	assert(str && *str);
 	assert(alloc_len);
 	assert(offset);
 	assert(src);
 
-	sz = strlen(src);
+	sz = (int)strlen(src);
 
 	if (*offset + sz >= *alloc_len) {
 		*alloc_len += sz < 32 ? 64 : 2 * sz;
